ATO-1389: add infrastructure for AuthUserInfo DynamoDb table
This table will replace the table of the same name that exists on Auth accounts. This table will store the UserInfo that Auth sents to Orch. Auth don't need to know about the table. In this migration, we also fix an issue with the current table, which is only indexed by internalCommonSubjectId. If a user has multiple client sessions, the authUserInfo entry gets overwritten. Hence, we have clientSessionId as a sort key as well. We keep ICSID as the partition key so we can still easily delete user data on request.