AUT-3910: Stop persisting unneeded generated state
`IPVReverificationService.buildIpvReverificationRedirectUri` currently stores the generated state that is sent to IPV for MFA reset against the session ID in Redis.

Since the cross-browser issue mitigation has been put in place `MfaResetAuthorizeHandler` also stores the same state in DynamoDB against client session ID.

We only need one of these. With the move away from Redis in progress it seems sensible to stick with DynamoDB.

The Redis version is also not read at the moment.