AUT-3910: Error if state returned from IPV is incorrect
When an MFA reset journey with IPV redirects the user back to auth, IPV will send back the state value auth sent to IPV when the journey started. We need to ensure that the state value received matches the state value sent for the users session. This helps prevent cross-site request forgery attacks.

We store that state value once generated inside the `ipv-reverification-state` DynamoDB table, alongside the client session ID. So access is given to the `ReverificationResultsHandler` to read from the table and a new error `ERROR_1061` is returned to the frontend if the states do not match.