AUT-3884: Replace @Required decorator with non-null check
The @Required decorator doesn't actually do anything. I tested by initialising an object with null values, and it worked!

To correctly disallow null values, you need to add a custom constructor to the record class to requireNonNull on the object.

The non-null check is only required for objects. Primitives naturally disallow null values.