BAU: Don't attach policy when orch stubbed
When orch is stubbed, the cross account access is not needed and
causes the deployment pipeline to fail in these environments.
This adds a conditional concat with the policy required policy
in to hopefully not attatch this unless required.

Again this lambda is not currently used in production so should
be fine to move this policy list around