ATO-1169: check if docAppSubjectId is defined in logs
This is the only place any setter on Auths ClientSession Class is called. Orch already sets it in AuthorisationHandler, so we should be able to safely remove this. If this assumoption is correct, then we can safely just pass the client session from orch to auth, and orch can forget about it. Ie, we don't need any info from Auth to set on the ClientSession.