BAU: Mirror images used in integration tests
The images we use in our integration tests can occasionally fail to pull
from docker hub.

Pull the images from docker hub and push them to the GitHub Container
Registry, to create our own mirror of the images.

Run the mirror job whenever the workflow, or the mirror job itself, is
updated. Also, run the mirror job once a week.