BAU: Make clearer we use rpPairwiseId for reauth counts
We use the rpPairwiseId to store reauth counts as well as the subjectId
from the user profile. This commit aims to clairfy this in the code
by using:

1 -  renaming the method params to make clearer we're using an optional
of the rpPairwiseId

2 - Clarify the test setup:

The unit test setup previously relied on mocking the configuration
service to return the same sector host as was setup for the
client, meaning that the value calculated for internalPairwiseId was
equal to the value calculated to the client, meaning we could reuse
the variable `expectedCommonSubject`, however this is not the case.

:wq