AUT-3976: Ensure that failure code in audit event is known
Introduces an enum to enforce that the failure_code we put into the audit event
is one of a known set of events, which matches what we've got in the event
schema.

If it is not, we log an error and omit from the audit event, but let the api return
a response to the frontend, which will handle returning an error to the user.