AUT-3974: Remove pairwise helper method
The new getRpPairwiseId method on the TxmaAuditHelper needs a change to
its call parameters to make it more generally callable.  Given that it
calls another helper and that changing the places it was called
would need changes that go beyond the scope of this ticket the decision
is to remove it.

This commit therefore removes the method and re-instates the call to
calculate the pairwise id back in the handler.  This resulted in the
unit and integration tests needing more setup code and checks.

The handler itself also had to be changed to add in a new constructor
allowing the Junit redis extension to be substituted for the Redis
service.