BAU: Introduce common test variables for account management tests
These variables are largely just needed to put headers onto a request, and are repeated in a lot of places. Pulling them out into somewhere common will reduce the noise in individual tests.

I haven't as part of this commit updated all the tests - this refactoring to follow in a subsequent commit or PR, just getting this in now to clean up my tests