AUT-4047: Add scripts used to create keys.
We needed to create a key from IPVs JWKS endpoint to add to our static
configuration as we haven't implemented the code that will retrieve it
automatically yet. The scripts to do this are included in this commit.

Additionally, we needed to create a key pair for our lower environment
deployments where we use the IPV stub.  The scripts used to generate
those keys are also included.