ATO-1470: add clientSessionId to the method interface
Because subjectId is a string, just changing the name and how it's used would not tell me if I've swapped over every use of this method to pass in clientSessionId instead of subjectId, as they are both strings. I've used intelisense and checking if it builds ok to see I have swapped over every use of this method. In a later commit, I will change it again to remove subjectId.