AUT-3484: Deployment of MFA Methods Retrieve
The openapi spec describes the endpoints with an operation id that takes
the format mfa-methods-<action>.  `mfa-methods` is the endpoint name and
`action` is the business name for what the User wants to do.

The operation id for retrieving an MFA Method is mfa-methods-retrieve.

The HTTP method currently being used to retrieve an MFA Method is GET as
per convention.  However, this could change to a POST and search pattern
in the future.  As GET is an implementation detail we have decided to
avoid using it and keep our names abstract.

To ensure that the terraform modules and the Java code are grouped
together in the project we will follow this style of naming.

For terraform this means all the resources created will either start
with or contain `mfa-methods-retrieve`.  The terraform file that
defines the lambda is therefore called `mfa-methods-retrieve.tf`.

The handler class is renamed to `MFAMethodsRetrieveHandler` to follow
this approach.