BAU: Add account verified param to setVerifiedPhoneNumber method
This method is currently only called in the case of account recovery. In account registration, we call a very similar method that sets the account verified status. We want to rationalise these two methods, and get rid of the account registration method. This parameter has therefore been introduced here for parity with the method called in account registration.

Note two things here:

* We are only ever setting account verified to true - this is the case for the calls to the existing registration method
* Account verified is not a flag used in any of our application logic - it was introduced to enabled some database analysis (see PR description here: https://github.com/govuk-one-login/authentication-api/pull/2232)