AUT-3848: Update dynamo service method to return optional user profile
This method is currently unused. In our new method management api, we'll want to look up a user profile from the public subject id. However, it's better for us to have this as an optional, rather than needing to catch an exception, so that we can 404. I've therefore updated the existing unused method to suit our purposes here.