AUT-3848: Introduce different mfa method records
The new records we originally introduced for the mfa methods work (SmsMfaData, AuthAppMfaData and MfaData) do not now reflect the structure of the mfa methods we want in the request and response for the method management api.

This commit introduces new data structures that we use on read from the database for method management. The intention is that these will replace the MfaData data structures that we originally introduced, but that will follow in a subsequent commit.

Introducing these means that we won't have to do another layer of transformation of the database object before we can return the object as a response