ATO-1522: Only pass previous session id if session present
There is a case where a user can log out after establishing a
session, and make another authorise request. When this happens
they have a session cookie but, no orchestration or shared session.
Logging out does not destroy a user's auth session.

Therefore if a user returns in this case, and we derive previous session id
just from the session cookie, it means a user will re-inherit an
auth session. This is not a problem, but is a divergence from the
way the shared session store worked, and means a user can end up
with fields on the auth session which are defined, but not on the
shared session.

To replicate the behaviour of the shared session we should only
pass previous session id to auth if the orch session exists.