AUT-2910: Record password reset success in session
When a user has successfully reset their password this is now recorded in the session. This will be used in a later commit to send the appropriate password reset payload to the TICF CRI API when the user finishes their journey.

An enum is used to record the `ResetPasswordState` rather than boolean as a later commit will add a `ATTEMPTED` state, as we also need to send to the TICF CRI API when an reset attempt was tried and the user failed to complete sign in.