AUT-3487: Return either in delete mfa method
Introduces the vavr library so that we can return values (in the form of eithers) rather than throwing exceptions that need to be caught to represent failure states.

An either represents a result that is either a left (indicating failure) or a right (indicating success). It means that the return values of functions can be values as opposed to throwing exceptions. We eventually would like to refactor the api to make this a standard pattern.