AUT-4158: Remove usages of auth-session-id and replace with session-id
The auth-session-id is the same as the regular session-id, so there's no need to log both.

Where we were only logging auth-session-id, I replaced it with a call to log session-id. Where we logged both, I removed the former.

Also, we weren't ever actually logging auth-session-id because it wasn't set correctly in the logging config.

This will make it easier to correlate logs across infrastructure, because we now have a single reference to session-id, shared across everything.