BAU: Rename mfa method constructors
These were specifically creating either an auth app mfa method or an sms mfa method, but it wasn't clear as we were just using the sugar for a constructor which doesn't differentiate on name.

This is a preparatory step for a refactor to change the interface of the addMfaMethodSupportingMultiple to take an MfaMethod rather than the now redundant MfaData object, as it'll be simpler to see what kind of mfa method we're constructing to pass into this method.

Co-authored-by: Aidan Comer <aidan.comer@digital.cabinet-office.gov.uk>