ATO-1117: add test for base64EncodedSalt in spot queue
the SpotRequest object holds salt as byte[], but in this ticket, I want to refactor it to use base64EncodedSalt instead. This is because the salt on the AuthUserInfo table is stored as a base64 string, and the way we send salt on to SPOT is also as a base64 string. I'm adding this test here to verify the raw string we send to Spot is the same thing now as it is with the new refactor. The one piece of nuance is that the `writeValueAsString` method changes the "=" to "\u003d". That happens currently, and this test proves that it will continue to happen when we pass in a string rather than a byte[].