AUT-2912: Send `2fa_method` property to TICF CRI
We have an existing `AuthSessionItem.verifiedMfaMethodType` property that is set when a new code is verified for `EMAIL`, `SMS` and `AUTH_APP` types for all journeys. Use this to send the value to the TICF CRI when it's `SMS` or `AUTH_APP` only. This works in conjunction with existing TICF CRI payload attributes (i.e. `initial_registration` and `2fa_reset`) + sign in journeys that are not explicitly marked in the TICF CRI payload.