AUT-2912: Introduce an internal representation for TICF CRI requests
The TICF CRI API specification is fairly non-standard, and we need to do unusual things like represent booleans as strings. We should capture all of this weirdness right as the edge, and represent requests in a way that matches our domain model everywhere else.

There are some additional challenges around lambda deserialising request objects: if the keys are not camel case then lambda will not deserialise by default and you need to provide a custom mapper. We can avoid all of that complexity by simply using a sensible object internally.