AUT-3486: Implement failure cases for updating default method
Current rules for updating a default method are that only the phone number or auth app credential can be changed. This commit implements all the invalid request cases as failures in the service, with a dummy value returned on success which will be updated subsequently.

Note that this will be used in a put request, so on a request to update the mfa method with no change, the actual result will be a success HTTP status (204) rather than a failure, but it will simplify the handler if implemented as a failure case here that the handler can handle differently.

As the update and create requests are the same shape, the name of the record has been updated.