AUT-4072: Create parameter for test client id
The acceptance tests need to use a client id that is available in all
environments and either has test clients enabled or that can be updated
by the tests as needed.  As the RP client records in the client-registry
that we want to use are created at deploy time by the TF we will also
populate the parameter store parameter at the same time.