ATO-1116: add test assertion to improve coverage
The case of a clientSession not existing for a clientId attached to the session was not tested and it was affecting coverage for this PR. There is no else, it is just whether we sent the backChannelLogout. If we can't find the clientSession, we can't do that, so this just asserts there was no attempt to.