ATO-1514: Add cross account combined policy
We're hitting a policy limit for the processing identity handler
when orch is not stubbed. This makes sense as we're exceeding 20
policies attatched to the role.

This creates a combined read, write, and delete role as well as a
combined KMS encrypt decrypt role for the CMK on the table.