AUT-3486: Return all of a user's mfa methods in update response
Following an adjustment to the schema, we now return all of a user's mfa methods after a request to update one of these. This makes sense since in some cases (not yet implemented) an update will change not just the mfa method being targeted by the identifier, but a user's other mfa method. This happens when switching default and priority methods.