AUT-3486: Return a 500 error when user has no default method
The only valid update on a backup is to switch it with a default method. We cannot do this if the user has no default method. This is being considered a 500 case as we it should not be possible for a user to get into this state and it signals an issue on our side.