AUT-3486: Add in failure cases for method to update all mfa methods for a user
When updating all methods, we should check that the combination of new methods is invalid.

This is only currently used in the put endpoint for mfa method management, and if any of these errors are returned we will return a 500 response from the endpoint. This is because these errors should not normally occur and, if they do, they will represent errors on our side