AUT-4123: Initial mfa property analysis lambda
We need to assess MFA property values across two DynamoDB tables in production to be confident we have not missed any combinations in our understanding before continuing part of the method management work.

Here this initial lambda will be tested in staging. It contains the logic for the more complex data fetch we'll need to do. After testing we can expand on what the lambda is doing with the data.

This lambda exists so we do not need to access any prod data directly. The lambda is expected to return the analysis results, not the source data itself.