AUT-2200: Refactor improve exception capture
Update the individual send methods to capture any exceptions that can
be thrown from the notificationService and log the details where they
occur.  Added explicit catches for checked exceptions that the compiler
warns about and a catch-all for any RuntimeExceptions that might be
raised.

Update the test to fix SonarQUBE issues for use of public keyword.