AUT-2200: Remove final exception handling
Applying the pattern of catching exceptions at source and not throwing
any exceptions to indicate failure to the rest of the lambda simplifies
it even more and allows us to remove the final tests confirming
exceptions were thrown.

The lambda does not allow exceptions to escape and so will not suffer
a restart when one of the classes it calls suffers an error that causes
it to throw an exception.