AUT-2200: Fix error logging and tests
The error messages were using %s instead of {} which is more efficient.
Introduced some constants for the personalisation values sent to Notify.

Increase test coverage to check all the errors are logged correctly.

Introduce a new class of constants for the messages logged by the
lambdas.  This allows the log messages to be tested without copying the
expected string content of the message.