AUT-4122: Ensure that migrated methods retain any existing mfa identifiers
Since we are now adding mfa identifiers on GET for non migrated methods, it's possible that a method might already have an mfa identifier when we migrate it, so here we ensure that it gets migrated across with the method.

Actually, based on the current implementation where we call getMfaMethodForNonMigratedUser first before migrating the user, we will currently always have an mfa identifier. However, some logic to generate an mfa identifier has been added in the case of migration to protect against changing this implementation