BAU: Create @Instrumented aspectj instrumentation
This allows us to use the @Instrumented annotation to automatically
instrument methods for telemetry. The instrumentation is done using
AspectJ, which allows us to weave in the instrumentation code at
compile time, load time, or runtime.

The end result is that we can add @Instrumented to any method, and spans
will be created, rather than having to wrap each method call with
`segmentedFunctionCall`.