AUT-4123: Count AUTH_APP enabled users with no verified MFA methods
One statistic we need to check is if there are any users that have enabled an AUTH_APP MFA method but with no verified AUTH_APP or SMS MFA methods.

This commits updates the query to only return the data we need to do that calculation from the user-credentials scan and user-profile batch get items.

A new class `MFAMethodAnalysis` is used to collate the statistic across forked batches and is returned by the handler as a string.