BAU: Move response_mode validation earlier
We introduced validation of the response_mode[1] previously, to align
with the OIDC spec[2] and handle an exception that was being thrown.

However, we added this to the end of the of the respective validation
which mean that when we redirected in any other invalid request error
cases we were not handling the invalid response mode. The spec says we
should return a Bad Request in any cases we don't understand this field,
and therefore we need to validate it first given it dictates __how__
we should return responses to the RP.

[1] - https://github.com/govuk-one-login/authentication-api/pull/6099
[2] - https://openid.net/specs/openid-connect-core-1_0.html#:~:text=If%20the%20Response%20Mode%20value%20is%20not%20supported%2C%20the%20Authorization%20Server%20returns%20an%20HTTP%20response%20code%20of%20400%20(Bad%20Request)%20without%20Error%20Response%20parameters%2C%20since%20understanding%20the%20Response%20Mode%20is%20necessary%20to%20know%20how%20to%20return%20those%20parameters