ATO-1514: Reset attempts when updating existing session
Our previously introduced consistency logs indicated these values
were out of sync some of the time. This was because we did not
reset the value when updating an existing orch session in the
same way we do in the `UpdateWithNewSessionId` method. This meant
that for users returning after an identity journey did not have this
value reset