AUT-2200: Allow NotificationHandler S3 access
Create the S3 bucket where the OTP codes can be stored for use in
acceptance testing.  The lambda that processes the events from the SQS
queue will need to know the name of the bucket to write the OTP to and
it will need the list of allow destinations (phone numbers and email
addresses) for which it should write to the bucket.