AUT-2200: Create S3 bucket for acceptance testing
The NotifyRequest sent by SendOtpNotificationHandler via the SQS event
needs a unique key when writing the OTP to S3 to ensure we don't get
contention for reading the values back in the acceptance tests which
are run multithreaded.  To ensure the S3 objects are unique for an
acceptance test we use the email address.  This is added to the
NotifyRequest event.

The NotificationHandler now checks test clients are allowed before
attempting to write to S3 instead of calling the Notify service to
ensure we always send to Notify in production where test clients are
disabled.

If the destination for the Notify request is on the allow list then
the NotificationHandler will now write the OTP to S3 using the email
as the key.