ATO-1515: Remove session resetting in start handler
Previously the start handler would "reset" the shared session if
a user came to it and orchestration considered them authenticated but
there was no user profile present. This was done in commit: 3a3752321071e6398b6ea7031bb58486d66b5016
to prevent a user deleting their profile in one tab, whilst authenticated
and then going on another journey where they were expected to go
on a silent login and encountering errors.

This behaviour no longer applies as we have done a small bit of
refactoring since (see commit: 0c20e56ce15813492d343123a61fa53b83aba633)
which means we check a user's profile exists and uses this as well
as the `authenticated` claim orchestration pass to auth to determine
whether we consider the user authenticated.

This session resetting behaviour is also becoming increasingly
redundant as we migrate the shared session store to two seperate
dynamo based stores.