AUT-2200: Add test to confirm ZDD
When the new version of the NotificationHandler is deployed to
production there could be a period of time when it is receiving SQS
events from the old version of the send OTP lambda which will not be
sending the new fields in the NotifyRequest.

Adding a test to check that the NotificationHandler can handle events
without the new fields and that it can also handle events with
unexpected fields in the event by ignoring them.