BAU: add dependabot config for terraform modules
This will keep all of our providers up to date without having to
manually check for updates.

As part of this, we stop using a single versions.tf, symlinked into each
component and instead switch to one individual file each. The symlink
approach was to avoid having to keep the versions up to date in each
component - dependabot now handles this for us.

This is safer & more granular: we can choose to configure only the
providers that are actually required by each component.

We're not allowing dependabot to update versions in modules: only the
'base' component versions.tf files, as modules will inherit the version
when loaded. Modules therefore have a lowest version constraint, which
we have tested the module at, rather than a specific required version.
Because of this, we don't need the lock files anymore.