AUT-4245: Only return AUT_APP method when verified
Following on from an investigation we have determined that we need to ensure only verified AUTH_APP MFA methods are returned. We had previously determined that we have no way to allow users to get into this state in our existing application (AUTH_APP records are only added after being verified).

However, an analysis of prod data with the MfaMethodAnalysisHandler shows we have some users in this state. Without this change those users could get locked out of their account.