AUT-4085: Refactor to clarify MFA models
The SmsMfaDetail class was being used for both request and response
mappings to the Method Management API.  We want to add otp checking
to MFA Handlers which callers of the API will provide in SmsMfaDetail
of the requests.

The otp is not stored in the database so when the details of an SMS
MFA Method are returned we need to return an SmsMfaDetail object but
without the otp field.

Various options were considered, such as continuing to use the same
class but making the otp optional so it could be used for both
purposes.  But after discussion we decided to create distinct classes
for the request and the response to avoid as much confusion as possible.

This commit introduces these new classes and paves the way for adding
otp checking in the next commit.