ATO-1596: get correctPairwiseId instead of rpPairwiseId
There are currently no clients that have a "public" subjectType that also go to identity, so this is not a bug. However, it is fragile, as if that were turned on, this would be the wrong value. I also want to make orchClientSession.getRpPairwiseId() to be private if possible so no dev uses it accidentally.

The log being removed here showed these two values are always the same.