BAU: Make DocApp JWKS implement beforeAll method
After moving the authorisation tests into nested classes I noticed that the
doc app tests were failing with ConnectionRefused when attempting
to fetch the JWKS.

When running the doc app tests individually they would pass, but
running the whole test suite they would fail. Having done some logging
I think the tests were attempting to spin up a new DocAppJwks extension
everytime a nested test class started but the port was in use. This defaulted
to the port -2 after the first nested test class finsihed (presumably because there
was one an extension running). This meant that when the DocAppJwksURI is constructed
it did not contain a port for  the local host url, resulting in a connection refused
exception when the doc app code attempted to fetch the JWKS.i

After some fiddling I found that making this implement the beforeAll method
seemed to solve this problem. I whave chosen this over modifying the base
httpStubExtension, as I am not sure if that would affect over tests