ATO-1596: get pairwiseId from orchClientSession instead of calculating
This makes use of the `getCorrectPairwiseIdGivenSubjectType`. There are currently no clients with the "public" subject type configured to for identity journeys. However, I want this code to be stable to a change to this, and I also want to make `getRpPairwiseId` on the orchClientSession private to protect devs from using it accidentally.