ATO-1515: Add Client Sessions to orch session when handling max_age
We previously re-assigned the client sessions from the previous
shared session to the current shared session. This does the same
for the orch session, in preperation for us swapping these fields
over.

There are also a couple of changes to make the test more realistic
and make the assertions more definitive, namely:
- Assert on the client session list directly
- Add a client session to the current session we setup