ATO-1515: Log if client session lists are equal
Adds a small logging util which logs if the client session lists
are the same. Have wrapped this in a try catch with a broad exception
catch as I don't want this to ever be able to cause an exception to be
thrown. Have added it to a fairly random smattering of handlers
but ones I believe should be representative of usages of this list
in destroy Session request.

Once we've confirmed these logs show they're in sync we should be able
to move over to using the Orch session for this value