BAU: Move cannot create default MFA rule
The rule that you cannot create a new DEFAULT MFA Method is only true
in account-management api.  The MFAMethodsService can be used to create
a new DEFAULT method, for example during sign-up.

Moving this check into the MFAMethodsCreateHandler enforces the rule
for updates to existing Users who must already have a DEFAULT method.