AUT-4205: Retrieve credential for migrated auth app user in auth app code processor
When a user's mfa methods have been migrated, we should retrieve the credential from a user's default mfa method, and error if any issues.

This involved adding a new error code, and using it for one of the existing scenarios (where a credential is not found, we return this new error instead of an 'invalid auth code' error, which makes more sense in this case anyway).

Some of the integration tests have been updated to reflect this different status code; another was not actually fully testing what it claimed to be, so has been updated to trigger the correct error message.