BAU: Rename and refactor test for bsid mismatch
This test was added as part of this commit: 037f9b20f2a94f04a2565b61a45c358be0821846
which added behaviour for a mismatch in bsid cookie and the session value.
The previous name was also inaccurate due to the fact that the request included
bsid cookie. I think this test was supposed to test the mismatch, and with subequent
changes has become a bit more inaccurate. This updates the tests to be a bit more
specific to the mismatch case.