BAU: Simplify API gateway event setup
We were manually setting up the API gateway event in a lot of tests
that were a bit unnecessary. We had a helper method in the file for
setting up the event for GET requests, so I have implemented this
more widely throughout all the tests.

I have also added an equivalent for POST requests and implmented where
appropriate