ATO-1596: remove logs that verify rpPairwiseId is in sync
Currently there are some handlers where rpPairwiseId is null on the orchClientSession. This looks like it is due to reading being too quick after writing. This is being fixed with strong consistent reads. Here however, these logs never returned false, as this is hit long enough after the orchClientSession is saved in authenticationCallbackHandler.