ATO-1116: remove temp logs to verify rpPairwiseId
There were 3 cases of rpPairwiseId not being in sync. The first was rpPairwiseId on the orchClientSession being null, as for that journey, they had not reached authenticationCallbackHandler. The second was for publicSubjectIds. The final is for rpPariwiseId being defined, but with a different value. We believe this is an existing bug that this actually patches, rather than a new bug being introduced. The given rpPairwiseId from Auth is the source of truth, not the one we calculate much later on in this service.