ATO-1116: only attempt to send backchannel logout if pairwiseId is defined
Currently the pairwiseId is calculated, which is defined if a user has signed in during a session at all. Now however, we get rpPairwiseId from the orchClientsession, which means it is only defined if a user reached authenticationCallbackHandler during that particular journey. We don't want to send a backchannel logout with a null pairwiseId, as the client won't know who to be signing out. Hence, add a check to see if pairwiseId is defined.

Here, also pull out the logic for checking if all the right values are defined, as with this extra check, we would have 3 nested ifPresent() callbacks, which was getting messy.