ATO-1596: add comment to not use these methods
These are only ever used interchangeably, ie we don't want one where we never want the other. We use one or the other always, as defined in the method getCorrectPairwiseIdGivenSubjectType. Unfortunately, we can't make them private, as dynamoDb requires the fields to be public for serialisation. We also can't change the name of the method. Hence, add this comment here.