AUT-4225: Add country code to number before storing on create backup mfa
This ensures that we store and compare numbers in a consistent format. Currently, we could add a backup of the same number but with a country code, and that would be considered a valid different number to the existing same number stored in the database without a country code.

This formatting is already done as standard in the frontend api when we store an sms method, so making sure the method management api is consistent with this.

Given we have not put the method management api live yet, we don't need to do any migration of non-normalised numbers since the only ones that exist are test numbers currently.

A subsequent commit will apply the same to the update endpoint