AUT-4225: Add country code to number before storing on update backup mfa
This ensures that we store and compare numbers in a consistent format.

This formatting is already done as standard in the frontend api when we store an sms method, so making sure the method management api is consistent with this.

Given we have not put the method management api live yet, we don't need to do any migration of non-normalised numbers since the only ones that exist are test numbers currently.