AUT-4202 - Move openapi spec to acc mgmt api
There are several versions of the Account Management OpenAPI spec.  This change makes
the spec in the terraform/account-management folder the reference version.  The copy
in the open-api specs repo will be removed.

A symlink makes the OpenAPI spec more accessible in the api-contract folder.  The
OpenAPI spec source file must remain in the terraform directories in order to pass
linting.

The Imposter configuration has been added to the api-contract folder to illustrate
how the API works.  There are sample httpie calls in the README.md.