BAU: Make getMfaMethods return a list of MFAMethods, not MfaMethodResponses
Part of making the MfaService care only about MFAMethods, not the response type which will be specific to account management (we are about to start using the MFAService in the frontend api as well).

Adds a method duplicated from the MfaService to convert mfa methods to mfa method responses to the retrieve handler. A subsequent commit will add this method to a shared account management helper, and further down the line once we've removed all traces of the MfaMethodResponse object from the MfaMethodsService, we can just remove the duplicated method from the MfaMethodsService

This involved updating a lot of the tests for other methods in the mfa methods service, which use the get method to check which methods remain stored after their operations.