BAU: Extract convertMfaMethodsToMfaMethodResponse to helper in account management
Extracts a method which will be reused in various endpoints of the new method management api to its own helper. Currently this method is also duplicated in the shared MfaMethodsService - we can remove that once all the public methods in the MfaMethodsService have been updated to only return MFAMethods, rather than the account-management specific MfaMethodResponse.

Currently this method returns a retrieve-specific failure - a subsequent commit will make it return a generic failure which can be used as appropriate in the account management handlers.

Part of a refactor to ensure that the MfaMethodsService is usable by the frontend api, as well as the account management api