BAU: Remove redundant inOrder calls
The InOrder class is to be used when a set of assertions needs
to be called in a specific order. All usages of this were single
assertions and therefore this was redundanBAU: Remove redundant inOrder calls

The InOrder class is to be used when a set of assertions needs
to be called in a specific order. All usages of this were single
assertions and therefore this was redundantt