ATO-1315: use single key access policy due to policy limit
Adding this new access policy to AuthenticationCallback will result in a policy limit error as this would exceed the quota for the role.

As we will be migrating away from using ipvTokenKeyArn anyway, I've merged the two policies. This policy will only apply to a single resource once the migration is complete.